Fix: A started measurement no longer blocks exit
The thread which makes the measurements continues even if the main
thread exits. This leads to a never ending program, if e.g. through an
error `measurement.end()` is not called. Now we make the thread a daemon
thread, so that it is automatically stopped once all non-daemon threads
exit.